# static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0",
      "title": "AutoCompany",
      "contact": {}
    },
    "host": "127.0.0.1:8000",
    "basePath": "/api",
    "securityDefinitions": {},
    "schemes": [
      "http"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
      "/products": {
        "get": {
          "summary": "Products",
          "tags": [
            "Misc"
          ],
          "operationId": "Products",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "required": true,
              "type": "integer",
              "format": "int32",
              "description": ""
            },
            {
              "name": "page_size",
              "in": "query",
              "required": true,
              "type": "integer",
              "format": "int32",
              "description": ""
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          }
        }
      },
...
    },
...
  }